refactor: 테스트 위치 올바른 위치로 이동
